Refactor: organize box controller and box routes
