--- src/controllers/boxController.js
@@ -1,288 +1,0 @@
-// const validator = require('validator');
-const boxDataMapper = require('../dataMappers/boxDataMapper');
-const { extractAndValidateBoxFields } = require('../utils/extractAndValidateBoxFields');
-const { generateSignedUrl } = require('../utils/signedUrl');
-const { getFromCache, setToCache } = require('../utils/cache');
-const { generateSignedUrlAndSaveItToCache } = require('../utils/generateSignedUrlAndSaveItToCache');
-
-const ttl = 60 * 60 * 24; // 86400 = 24 hours
-
-// ----- getBoxById -----
-// Voir commentaires dans getBoxes
-// TODO : faire la doc endpoint et code erreur
-const getBoxById = async (req, res) => {
-  try {
-    const { boxId } = req.params;
-    const box = await boxDataMapper.getBoxById(boxId);
-    if (!box) {
-      return res.status(404).json([{ errCode: 29, errMessage: 'Box not found' }]);
-    }
-
-    if (box.picture && box.picture.pictureUrl) {
-      const s3Url = box.picture.pictureUrl;
-      const s3ObjectKey = s3Url.split('/').pop();
-      const cacheKey = `box:${box.id}:picture`;
-      const cachedUrl = await getFromCache(cacheKey);
-      if (cachedUrl) {
-        box.picture.pictureUrl = cachedUrl;
-      } else {
-        const signedUrl = await generateSignedUrl(s3ObjectKey, ttl);
-        await setToCache(cacheKey, signedUrl, ttl);
-        box.picture.pictureUrl = signedUrl;
-      }
-    } else {
-      box.picture = {
-        pictureUrl: '',
-        photographerName: '',
-        photographerProfileUrl: '',
-      }; // renvoie cet objet plutôt que null pour éviter les erreurs côté client
-    }
-
-    return res.status(200).json(box);
-  } catch (error) {
-    console.error({ getBoxByIdError: error });
-    return res.status(500).json([{ errCode: 30, errMessage: 'A server error occurred when retrieving the box' }]);
-  }
-};
-
-// ----- getBoxes -----
-const getBoxes = async (req, res) => {
-  try {
-    // Dans le middleware authenticateToken, on a ajouté les infos utilisateur à l'objet req
-    const userId = req.user;
-    const boxes = await boxDataMapper.getBoxes(userId);
-
-    // Si il y a des boxes, on va vérifier pour chacune d'elles si elles ont une image.
-    // Si oui : on vérifie si un lien signé est présent en cache :
-    // +++ si oui, c'est bien on la récupère ; +++ si non, on génère un lien signé, on le sauvegarde en cache, on le renvoie avec sa box
-    // Si non : on ne touche pas, on renvoie la box telle quelle avec box_picture = null
-    // Pour des opérations asynchrone sur réalisé sur un objet itérable, on va utiliser Promise.all()
-
-    if (boxes.length > 0) {
-      // console.log('----- Il y a des boxes, je vais vérifier si elles ont une image -----');
-      const boxPromises = boxes.map(async (box) => {
-        // console.log(box.name, " : je travaille dans cette box et vérifie la présence d'une image");
-        if (box.picture && box.picture.pictureUrl) {
-          const s3Url = box.picture.pictureUrl;
-          const s3ObjectKey = s3Url.split('/').pop(); // la clé de l'objet dans S3 (en fait, le nom du fichier présent dans l'URL)
-          // console.log(box.name, " : Oui, il y a une image, je vérifie dans redis si l'url signée est en cache");
-          const cacheKey = `box:${box.id}:picture`;
-          const cachedUrl = await getFromCache(cacheKey);
-          if (cachedUrl) {
-            // console.log(box.name, ' : Oui, elle y est, cachedUrl est renseignée et la box est renvoyée');
-            return {
-              ...box,
-              picture: {
-                pictureUrl: cachedUrl,
-                photographerName: box.picture.photographerName,
-                photographerProfileUrl: box.picture.photographerProfileUrl,
-              },
-            };
-          }
-          // console.log(
-          //   box.name,
-          //   " : Non, elle n'y est pas, je génère un lien signé, je le sauvegarde en cache, je le renvoie avec sa box"
-          // );
-          const signedUrl = await generateSignedUrl(s3ObjectKey, ttl);
-          await setToCache(cacheKey, signedUrl, ttl);
-          return {
-            ...box,
-            picture: {
-              pictureUrl: signedUrl,
-              photographerName: box.picture.photographerName,
-              photographerProfileUrl: box.picture.photographerProfileUrl,
-            },
-          };
-        }
-        // console.log(box.name, " : Non, il n'y a pas d'image, je renvoie la box telle quelle avec picture_url = null");
-        return {
-          ...box,
-          picture: {
-            pictureUrl: '',
-            photographerName: '',
-            photographerProfileUrl: '',
-          }, // renvoie cet objet plutôt que null pour éviter les erreurs côté client };
-        };
-      });
-      const boxesWithSignedUrls = await Promise.all(boxPromises);
-      // Si boxes : on renvoie les boxes avec les liens signés
-      return res.status(200).json(boxesWithSignedUrls);
-    }
-    // Si pas de boxes : on renvoie un tableau vide
-    return res.status(200).json([]);
-  } catch (error) {
-    console.error({ getBoxesError: error });
-    return res.status(500).json([{ errCode: 31, errMessage: 'A server error occurred when retrieving the boxes' }]);
-  }
-};
-
-// ----- createBox -----
-const createBox = async (req, res) => {
-  try {
-    const fields = extractAndValidateBoxFields(req);
-    if (fields.error) {
-      return res.status(400).json([fields.error]);
-    }
-
-    const {
-      userId,
-      name,
-      description,
-      color,
-      label,
-      level,
-      defaultQuestionLanguage,
-      defaultQuestionVoice,
-      defaultAnswerLanguage,
-      defaultAnswerVoice,
-      learnIt,
-      type,
-      pictureUrl,
-      photographerName,
-      photographerProfileUrl,
-    } = fields;
-
-    const createdBox = await boxDataMapper.createBox(
-      userId,
-      name,
-      description,
-      color,
-      label,
-      level,
-      defaultQuestionLanguage,
-      defaultQuestionVoice,
-      defaultAnswerLanguage,
-      defaultAnswerVoice,
-      learnIt,
-      type,
-      pictureUrl,
-      photographerName,
-      photographerProfileUrl
-    );
-    // Generate signed URL for box picture and save it to cache
-    if (createdBox.picture && createdBox.picture.pictureUrl) {
-      // Replace the box_picture with a signed URL and save signed URL to cache
-      createdBox.picture.pictureUrl = await generateSignedUrlAndSaveItToCache(
-        { userOrBoxOrCard: 'box', id: createdBox.id, infoType: 'picture' },
-        createdBox.picture.pictureUrl,
-        ttl
-      );
-    } else {
-      createdBox.picture = {
-        pictureUrl: '',
-        photographerName: '',
-        photographerProfileUrl: '',
-      }; // renvoie cet objet plutôt que null pour éviter les erreurs côté client
-    }
-
-    return res.status(201).json(createdBox);
-  } catch (error) {
-    console.error({ createBoxError: error });
-    return res.status(500).json([{ errCode: 32, errMessage: 'A server error occurred when creating the box' }]);
-  }
-};
-
-// ----- updateBox -----
-const updateBox = async (req, res) => {
-  try {
-    const { boxId } = req.params;
-    const fields = extractAndValidateBoxFields(req);
-    if (fields.error) {
-      return res.status(400).json([fields.error]);
-    }
-    const {
-      name,
-      description,
-      color,
-      label,
-      level,
-      defaultQuestionLanguage,
-      defaultQuestionVoice,
-      defaultAnswerLanguage,
-      defaultAnswerVoice,
-      learnIt,
-      type,
-      photographerName,
-      photographerProfileUrl,
-    } = fields;
-
-    let { pictureUrl } = fields;
-    if (!pictureUrl && req.body.existingImageUrl) {
-      // If no new image is uploaded, we keep the existing image
-      pictureUrl = req.body.existingImageUrl;
-    }
-
-    const updatedBox = await boxDataMapper.updateBox(boxId, {
-      name,
-      description,
-      color,
-      label,
-      level,
-      defaultQuestionLanguage,
-      defaultQuestionVoice,
-      defaultAnswerLanguage,
-      defaultAnswerVoice,
-      learnIt,
-      type,
-      pictureUrl,
-      photographerName,
-      photographerProfileUrl,
-    });
-    // Generate signed URL for box picture and save it to cache
-    if (updatedBox.picture && updatedBox.picture.pictureUrl) {
-      // Replace the box_picture with a signed URL and save signed URL to cache (key: box:${boxId}:image)
-      updatedBox.picture.pictureUrl = await generateSignedUrlAndSaveItToCache(
-        { userOrBoxOrCard: 'box', id: updatedBox.id, infoType: 'picture' },
-        updatedBox.picture.pictureUrl,
-        ttl
-      );
-    } else {
-      updatedBox.picture.pictureUrl = '';
-    }
-    return res.status(200).json(updatedBox);
-  } catch (error) {
-    console.error({ updateBoxError: error });
-    return res.status(500).json([{ errCode: 132, errMessage: 'A server error occurred when updating the box' }]);
-  }
-};
-
-// ----- updateBoxLearnItValue -----
-const updateBoxLearnItValue = async (req, res) => {
-  try {
-    const { boxId } = req.params;
-    const { learnIt } = req.body;
-    if (typeof learnIt !== 'boolean') {
-      return res.status(400).json([{ errCode: 41, errMessage: 'Invalid learnIt value' }]);
-    }
-    const newState = await boxDataMapper.updateBoxLearnItValue(boxId, learnIt);
-    return res.status(200).json(newState);
-  } catch (error) {
-    console.error({ updateBoxLearnItValueError: error });
-    return res
-      .status(500)
-      .json([{ errCode: 42, errMessage: 'A server error occurred when updating "learn it" state' }]);
-  }
-};
-
-// ----- deleteBox -----
-const deleteBox = async (req, res) => {
-  try {
-    const { boxId } = req.params;
-    await boxDataMapper.deleteBox(boxId);
-    return res.status(204).end();
-  } catch (error) {
-    console.error({ deleteBoxError: error });
-    return res.status(500).json([{ errCode: 40, errMessage: 'A server error occurred when deleting the box' }]);
-  }
-};
-
-const boxController = {
-  getBoxById,
-  getBoxes,
-  createBox,
-  updateBox,
-  updateBoxLearnItValue,
-  deleteBox,
-};
-
-module.exports = boxController;
