+++ src/controllers/boxController/index.js
@@ -1,0 +1,15 @@
+const getBoxById = require('./getBoxById');
+const getBoxes = require('./getBoxes');
+const createBox = require('./createBox');
+const updateBox = require('./updateBox');
+const updateBoxLearnItValue = require('./updateBoxLearnItValue');
+const deleteBox = require('./deleteBox');
+
+module.exports = {
+  getBoxById,
+  getBoxes,
+  createBox,
+  updateBox,
+  updateBoxLearnItValue,
+  deleteBox,
+};
