+++ src/routes/index.js
@@ -1,0 +1,8 @@
+const { Router } = require('express');
+const boxRoutes = require('./boxRoutes');
+
+const router = Router();
+
+router.use('/box', boxRoutes);
+
+module.exports = router;
